main/requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.4
Pillow==5.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
reverse==0.1.0
serializers==0.2.4
sqlparse==0.3.1
virtualenv==16.7.9
whitenoise==3.3.1
